Fuck Vue files with a god damn hammer
Also fuck webpack.
Also fuck web dev.
Also fuck Javascript.
Also fuck everything.

Fuck all this noise.
