+++ package.json
@@ -6,8 +6,8 @@
   "private": true,
   "scripts": {
     "build": "node build/build.js",
-    "dev": "node build/dev-server.js",
-    "lint": "tslint -c tslint.json 'src/**/*.ts'",
+    "dev": "webpack-dev-server --no-inline --progress --config build/webpack.dev.conf.js",
+    "lint": "tslint -p tsconfig.json -c tslint.json 'src/**/*.ts'",
     "start": "npm run dev",
     "test": "yarn run unit",
     "unit": "BABEL_ENV=test karma start test/unit/karma.conf.js --single-run"
@@ -24,6 +24,7 @@
     "autoprefixer": "^7.1.2",
     "babel-core": "^6.22.1",
     "babel-loader": "^7.1.1",
+    "babel-plugin-istanbul": "^4.1.5",
     "babel-plugin-syntax-async-functions": "^6.13.0",
     "babel-plugin-transform-regenerator": "^6.26.0",
     "babel-plugin-transform-runtime": "^6.22.0",
@@ -32,15 +33,17 @@
     "babel-preset-stage-2": "^6.22.0",
     "babel-register": "^6.22.0",
     "chalk": "^2.0.1",
-    "connect-history-api-fallback": "^1.3.0",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.0",
-    "express": "^4.14.1",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^1.1.4",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "file-loader": "^1.1.11",
     "friendly-errors-webpack-plugin": "^1.6.1",
+    "glob": "^7.1.2",
+    "gulp": "^4.0.0",
+    "gulp-cli": "^2.0.1",
+    "gulp-tslint": "^8.1.3",
+    "html-loader": "^0.5.5",
     "html-webpack-plugin": "^2.30.1",
-    "http-proxy-middleware": "^0.17.3",
     "jasmine": "^3.1.0",
     "karma": "^2.0.0",
     "karma-coverage": "^1.1.1",
@@ -50,17 +53,22 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "^0.0.32",
     "karma-webpack": "^2.0.13",
-    "opn": "^5.1.0",
+    "node-notifier": "^5.2.1",
+    "node-sass": "^4.7.2",
     "optimize-css-assets-webpack-plugin": "^3.2.0",
     "ora": "^1.2.0",
     "phantomjs-prebuilt": "^2.1.16",
     "portfinder": "^1.0.13",
     "rimraf": "^2.6.0",
+    "sass-loader": "^6.0.7",
     "semver": "^5.3.0",
     "shelljs": "^0.7.6",
+    "style-loader": "^0.20.3",
     "ts-loader": "^3.5.0",
+    "tsconfig-paths-webpack-plugin": "^2.0.0",
     "tslint": "^5.9.1",
     "tslint-config-airbnb": "^5.7.0",
+    "tslint-consistent-codestyle": "^1.12.2",
     "tslint-loader": "^3.6.0",
     "typescript": "^2.7.2",
     "url-loader": "^0.5.8",
@@ -69,7 +77,8 @@
     "vue-template-compiler": "^2.5.0",
     "webpack": "^3.6.0",
     "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-middleware": "^1.12.0",
+    "webpack-cli": "^2.0.13",
+    "webpack-dev-server": "^2.0.0",
     "webpack-merge": "^4.1.0"
   },
   "engines": {
