+++ src/app.html
@@ -1,0 +1,6 @@
+<div id="app" class="full-height">
+  <nav-pane :tabs="tabs" :selected.sync="selected" v-on:select="handleSelect">
+    <router-view>
+    </router-view>
+  </nav-pane>
+</div>
