+++ src/app.scss
@@ -1,0 +1,22 @@
+html {
+  height: 100%;
+}
+
+body {
+  height: 100%;
+  margin: 0px;
+  background-color: #151515;
+  color: #DDDDDD;
+}
+
+.full-height {
+  height: 100%;
+}
+
+a {
+  color: #DEDEDE;
+}
+
+a:visited {
+  color: #A6A6A6;
+}
