--- src/components/TopBar.vue
@@ -1,29 +1,0 @@
-<template>
-  <div class="main-header">
-    <div class="main-header-menu">
-      <slot>
-      </slot>
-    </div>
-  </div>
-</template>
-
-<script lang="ts">
-export default {
-  name: 'TopBar',
-};
-</script>
-
-<style scoped>
-.main-header {
-  height: 40px;
-  background-color: #222222
-}
-
-.main-header-menu {
-  float: right;
-  margin-right: 5px;
-  top: 50%;
-  transform: translateY(-50%);
-  position: relative;
-}
-</style>
