+++ src/components/add-button.html
@@ -1,0 +1,1 @@
+<button class="add-button">+</button>
