+++ src/components/add-button.scss
@@ -1,14 +1,3 @@
-<template>
-  <button class="add-button">+</button>
-</template>
-
-<script lang="ts">
-export default {
-  name: 'AddButton',
-};
-</script>
-
-<style>
 .add-button {
   width: 20px;
   height: 20px;
@@ -26,4 +15,3 @@ export default {
 .add-button::-moz-focus-inner {
   border-style: none;
 }
-</style>
