+++ src/components/add-button.ts
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+
+import template from './add-button.html';
+
+export const AddButton = Vue.component('add-button', {
+  name: 'AddButton',
+  template,
+});
