+++ src/components/nav-pane.scss
@@ -1,0 +1,25 @@
+.left-bar {
+  background-color: #222222;
+  float: left;
+  height: 100%;
+  width: 100px;
+}
+
+.left-button {
+  padding-left: 10px;
+  padding-top: 7px;
+  padding-bottom: 7px;
+}
+
+.left-button:hover {
+  background-color: #404040;
+  cursor: pointer;
+}
+
+.content-pane {
+  margin-left: 100px;
+}
+
+.selected {
+  background-color: #303030;
+}
