+++ src/components/top-bar.html
@@ -1,0 +1,6 @@
+<div class="main-header">
+  <div class="main-header-menu">
+    <slot>
+    </slot>
+  </div>
+</div>
