+++ src/components/top-bar.js
@@ -1,0 +1,8 @@
+"use strict";
+exports.__esModule = true;
+var vue_1 = require("vue");
+var top_bar_html_1 = require("./top-bar.html");
+exports.TopBar = vue_1["default"].component('top-bar', {
+    name: 'TopBar',
+    template: top_bar_html_1["default"]
+});
