+++ src/components/top-bar.scss
@@ -1,0 +1,12 @@
+.main-header {
+  height: 40px;
+  background-color: #222222
+}
+
+.main-header-menu {
+  float: right;
+  margin-right: 5px;
+  top: 50%;
+  transform: translateY(-50%);
+  position: relative;
+}
