+++ src/components/top-bar.ts
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+
+import template from './top-bar.html';
+
+export const TopBar = Vue.component('top-bar', {
+  name: 'TopBar',
+  template,
+});
