+++ src/components/video-list-component.html
@@ -1,0 +1,25 @@
+<div id="videoList">
+  <top-bar>
+    <add-button v-on:click.native="showNewVideo"></add-button>
+  </top-bar>
+
+  <div v-if="loading">
+    Loading...
+  </div>
+
+  <div v-if="error">
+    Error: {{ error }}
+  </div>
+
+  <div v-for="video in videos">
+    <router-link :to="makeVideoUrl(video.id)">{{ video.name }}</router-link>
+  </div>
+
+  <input
+    v-if="this.addVideo"
+    ref="addVideoInput"
+    v-model="newVideo"
+    v-on:keyup.enter="createVideo"
+    v-on:blur="createVideo">
+  </input>
+</div>
