+++ src/components/video-list.scss
@@ -1,0 +1,3 @@
+#videoList {
+  text-align: center;
+}
