+++ src/components/video-player.html
@@ -1,0 +1,5 @@
+<div>
+  <video id="video" preload="metadata" controls>
+    <source :src="$route.params.fileUrl">
+  </video>
+</div>
