+++ src/components/video-player.scss
@@ -1,0 +1,5 @@
+#video {
+  position: absolute;
+  height: 100%;
+  width: 100%;
+}
