+++ src/components/video-player.ts
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+
+import template from './video-player.html';
+
+export const VideoPlayer = Vue.component('video-player', {
+  name: 'VideoPlayer',
+  template,
+});
