+++ src/components/video.scss
@@ -1,0 +1,3 @@
+#videoFiles {
+  text-align: center;
+}
