+++ src/html-shim.ts
@@ -1,0 +1,4 @@
+declare module '*.html' {
+  const content: string;
+  export default content;
+}
