+++ src/main.ts
@@ -1,12 +1,14 @@
 import Vue from 'vue';
-import App from './App';
+import { App } from './app';
 import router from './router';
+
+import './styles.scss';
 
 Vue.config.productionTip = false;
 
 new Vue({
   el: '#app',
   router,
-  template: '<App/>',
+  template: '<app/>',
   components: { App },
 });
