+++ src/services/video.spec.ts
@@ -1,8 +1,6 @@
-describe('foo', function() {
-  beforeAll(function() {
-    console.log('hi');
-  });
+import { Video } from './video';
 
+describe('video service wrapper', function() {
   it('should work', function() {
     expect(true).toBeTruthy();
     expect(false).toBeFalsy();
