+++ src/styles.scss
@@ -1,0 +1,9 @@
+@import './app';
+
+@import './components/add-button';
+@import './components/library-list';
+@import './components/nav-pane';
+@import './components/top-bar';
+@import './components/video';
+@import './components/video-list';
+@import './components/video-player';
