+++ tsconfig.json
@@ -7,7 +7,12 @@
 		"noImplicitAny" : true,
 		"noImplicitReturns" : true,
     "noImplicitThis": false,
-    "lib": [ "es2015", "dom" ]
+    "lib": [ "es2015", "dom" ],
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"],
+      "@components/*": ["src/components/*"]
+    }
   },
   "include": [
     "./src/**/*"
